+++ src/index.css
@@ -92,6 +92,7 @@ img {
 
 .oblong{
   display: inline-block;
+  cursor: pointer;
 }
 
 .oblong:active{
@@ -235,7 +236,8 @@ img {
   top: 32%;
   width: 100%;
   height: 300px;
-  display: inline-block;
+  /* display: inline-block; */
+  white-space: nowrap;
 }
 
 .scoreboard:hover{
@@ -243,7 +245,8 @@ img {
 }
 
 .scoreboard{
-  float:left;
+  /* float:left; */
+  display: inline-block;
   position: relative;
   top: 0px;
   height: 300px;
