+++ src/pages/Nban.js
@@ -24,7 +24,7 @@ function Nban(props){
 
         <Link to="/alongwalk" style={{position: 'absolute', top: '150px', left: '900px', color: '#083182'}}>a long walk to somewhere close</Link>
 
-        <Link to="/slippery" style={{position: 'absolute', width: '100px', top: '550px', left: '100px', color: '#083182'}}>slippery</Link>
+        <Link to="/following-strangers" style={{position: 'absolute', width: '100px', top: '550px', left: '100px', color: '#083182'}}>following strangers</Link>
 
     </div>
   );
